misc/celery/celery_requirement.txt
amqp==5.0.1
billiard==3.6.3.0
celery==5.0.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
vine==5.0.0
redis==3.5.3
six==1.15.0
psycopg2-binary==2.8.6
wcwidth==0.2.5
prompt-toolkit==3.0.8
kombu==5.0.2
pytz==2020.1
